Add `accumulator` crate
Creates an `Accumulator` structure that contains high and low bits.
Allows adding and spending an `OutPoint` from the set.

The `OutPoint` is serialized to component parts and concatinated. A
potential speed up is to provide a 32 byte input to the tagged hash,
which could potentially speed up the hash step by omitting a 32 byte
allocation.

Internally the addition first does a `wrapping_add` on the high bits, a
`overflowing_add` on the low bits, and does a carry if there was an
overflow. The subtraction performs an `add` on the inverse of the value.