Update accumulator and add a UTXO example
Ruben mentioned there is no optimization for using a 32 byte input to
the tagged hash, so that change is unrolled here. This also adds an
example that spends and adds from the accumulator given a table called
`outpoints.db`. To build the table one would run `utxo-to-sqlite.py` in
`bitcoin` on a `utxos.dat` file.