verifier: Remove shared lock on the accumulator
The shared lock on the accumulator between the block download tasks
and the blocking database task that loads UTXOs from sqlite is a
performance drag.

This refactors the verification assumption to say the accumulator state
of the UTXO set should be the same as the accumulator state after doing
block download to the assume valid hash.