verifier: Switch to `std::net` impl
The `tokio` framework requires additional context outside of the
standard Rust knowledge. Now there are `WORKERS` number of threads that
are continually fetching blocks, and a dedicated hashing thread that is
updating an accumulator. Simultaneously, there is an accumulator updated
from the outpoints from the UTXO set. Once every worker thread
completes, the task handles are joined and the accumulator states are
compared.